Rationalise variants:fromX tags and just support Statamic filtering
